pyproject.toml
[tool.poetry]
name = "speedrun"
version = "0.1.0"
description = ""
authors = ["theoremoon <theoldmoon0602@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
